+++ lib/features/financial/shared/constants/financial_constants.dart
@@ -1,0 +1,5 @@
+class FinancialConstants {
+  static const double minimumLoanAmount = 10000.0;
+  static const double maximumLoanAmount = 5000000.0;
+  static const double minimumInvestmentAmount = 500.0;
+}